requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-cors==1.0.0

# Environment Variables
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2

# OTP Generation
pyotp==2.9.0

# SMS Service (Twilio)
twilio==8.11.1

# Email Service
python-smtp==1.0.0

# Utilities
python-slugify==8.0.1

pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.9
